+++ apps/api/src/ee/services/webhook/webhook.interfaces.ts
@@ -1,4 +1,4 @@
-import { RabbitMQAlert } from "@/services/alerts/alert.interfaces";
+import { RabbitMQAlert } from "@/ee/services/alerts/alert.interfaces";
 
 export interface WebhookPayload {
   version: string;
