+++ apps/api/src/trpc/routers/rabbitmq/index.ts
@@ -1,6 +1,7 @@
 import { router } from "@/trpc/trpc";
 
-import { alertsRouter } from "./alerts";
+import { alertsRouter } from "../../../ee/routers/rabbitmq/alerts";
+import { topologyRouter } from "../../../ee/routers/rabbitmq/topology";
 import { definitionsRouter } from "./definitions";
 import { infrastructureRouter } from "./infrastructure";
 import { memoryRouter } from "./memory";
@@ -10,7 +11,6 @@ import { overviewRouter } from "./overview";
 import { policiesRouter } from "./policies";
 import { queuesRouter } from "./queues";
 import { serverRouter } from "./server";
-import { topologyRouter } from "./topology";
 import { usersRouter } from "./users";
 import { vhostRouter } from "./vhost";
